Added a new wexpr constructor (W_UPDATE_NAMED_SLICE) to make the visualization of the common slice update function significantly better and the simulation faster.
Fixed the visualization of predicates to include the inputs.

Improved the visualization of tuple assignments.

Improved the visualization of finite state machines so that text is scaled
as the image is scaled (zoomed).

Improved the visualization of code that uses structs in which some signals
are inputs and others are outputs.