Fixed a major bug in the pexlif2fsm function triggered when using lists of inputs with non-contiguous vectors.
Removed the overloading of BDD and bexpr for the antecedent DSL. It just
added complexity for something rarely (never?) used.

Enhanced the DSL for antecedent/consequents to allow X (unconstrained) as
value to make sequences easier to write.

Fixed a number of small bugs/poor features in HFL.

Fixed the error reporting for sprintf like functions when an integer
does not fit in the width given. Previously it lost the error message
which made debugging very difficult!

Made a major change to open-overload resolution that always tries to
resolve the overloading before adding "hidden" arguments.
This fixes the long standing problem with loss of laziness for expressions
that uses open overloaded functions.
The main problem is that you must now ensure there are at least two
alternatives overloaded before defining functions that you want to be
implicitly overloaded.  This fix is an experiment and we have to see
if it is viable.

Fixed a silly bug in the input/list_input functions that made mixed inputs
and outputs with expressions sometime misname arguments. Trivial fix, but
it took almost a week to find the problem..... Yikes!

Added inout as a binder for structures that have some signals used as inputs
and others as outputs.

Added a circuit that takes an input (vector) and outputs a 1 only after the
input has been the same for duration_in_phases phases.

Fixed the drawing functions for "fubs" with no inputs.