Performance improvement in hfl.
Fixed a bad race condition in the user interface (tcl_eval calling a tcl
function that called fl back that called a tcl_eval etc.).

Added a monitor construct to allow custom-written value display linked to
an STE run.

Provided a means to display only a subset of nodes (useful for memories!).

Fixed the Time window in the waveform and schematcs visualization to only
take effect once the Return key is pressed.

Added a command md_sizes that finds the total size of a list of vectors.

Fixed several bugs in merge_vectors and md_merge_vectors.

Enhanced the display_dot tcl procedure to allwo a window to be reused and also
to provide a cancel function when the window is closed.