Cleaned up the Verilog writer to use the new version of yosys and to use gui_exec.
Modified some of the examples to use idv.fl instead of ste.fl.

Made reading .vossrc files printing out an information mesasge
so that the user knows what is going on.

Enhanced the assertion function to make it much faster by using
pre-computed results.

Fixed Get_Vector_Size to accept x's in addition to 0's and 1's.

Fixed a silly typo in write_pexlif.cc (arguments to _adff was in wrong order).