Enabled SX/ZX and bit-selection from wexpr types (i.e., of size 0 type). This allows generic models to be built using SX/ZX and bit-selection.
Made flfun2pexlif a free_binder to make it much easier to instantiate
top-level HFL fubs.

Fixed a missing non_lazy for the redefintions of tcl_eval etc.

Started the refactoring of the idv_gui code towards separating the
gui parts from the API part.

Fixed the help system to work better when (re-)attaching the help window
from the main VossII window.

Fixed a major bug in pexlif2fsm when when non-contiguous vectors were found.
The code was basically completely broken and needed a re-write.

Fixed the code that generates fresh variables to also look at the
inputs and outputs lists.

Fixed a bug in constant generation.

Removed duplicated code that had snuk in after a cut-and-paste that
made VossII run out of file descriptors (since they never got closed!).

Fixed the cleanup code for making PINSTs to remove duplicat SHA and FP
arguments in the atribute list.