Improved the creation of LEAF cells so that they all have anonymous wirenames and thus the same SHA256 signatures. In the process, cleaned up the split code in hfl.fl.
Changed the interface to the w_substitute function so that multi-dimensional
wires can be substituted.

Added loading vector_names.fl in util.fl since it now uses vec_assoc.

Fixed a nasty bug in Graph_hash that computed different hash signatures for
a DAG and the corresponding tree.... Example of correct, but slow code, that
changed three times, each looking like an ok thing to do but together made it
misbehave....

Added set_eq and set_neq as builtins since that is significantly faster (less
garbage produced).

Added a Info menu to provide more low-level details of a drawn object.

Fixed the "i_am_busy" / "i_am _free" so that the GUI does not freeze.

Started creating a clean API to the transformations. Still WIP.

Fixed the unfold command to work correctly when multiple instances are selected.

Did a major overhaul of the replace/replace-by-patter/ etc. code to work
much more predcictably and quickly. I suspect more work will be needed
on performance tuning.

Fixed the Verilog writer for W_EXPLICIT_VAR constructs.

Added an exception handler in GUI mode for recursion limit reached. Now the
user is given a choice to continue with the limit doubled rather than just
raising an exception.