Fixed a bug in the fl function const_prop that handled SLICES incorrectly.
Added a constant propagation transformation. Still WIP.

Fixed the Makefile so that lex.yy.c and y.tab.c is compiled with the
same arguments as all the other C files.

Added error checking in the builtin functions gather and separate to ensure
the selection indices are within the range.

Fixed the definition of zip3, zip4, etc. to allow lists of different types.
Duh.....