Added some new GUI functionality for the waveform viwer (can now save/load the curent waveform nodes shown).
Fixed a major bug in the counterexample generation code that sometimes
did not genereate a counter example (even though the verification established
the presence of such example).

Fixed a bug in the graph reduction engine that made the failure message
sometimes get lost and only a "Failure:" ended up being shown.