Fixed a number of Bifrost functions to use the new bundle construct.
Added a function to create a interconnect pexlif to compine a number
of pexlif functions.

Removed pcompile for functions with bundle interfaces since pcompile cannot
handle them at the moment (pcompile might never handle them).

Found a bug in simplify_W_AND so commented it out for the time being.
I suspect it is the wexpr2bool function that is incorrect (not computing
the wexpr size correctly). WIP!

Fixed a nast typo in list_outputs that made some signals inputs....

Fixed a bug in the constant propagation code.

Changed the destination for warnings from stderr to stdout in -noX mode.
Now only real errors go to stderr.

Fixed a bug that failed to close an open file handler for a particular
error condition.

Fixed a subtle bug in the non_lazy (and default argument insertion) to
only orocess a node once.

Fixed an ugliness in the GUI that made it hard to read some text.