Made bifrost clockgating logic more parameterized.
Fixed a bug in w2bv that translated negative constants incorrectly.

Minor GUI enhancements to allow keyboard traversal among notebook tabs.

Added the option -- to fl that stops parsing for options. Anything after --
will be prvided in ARGS in fl.

Added flushing after a stdinfo printf so that information is not buffered.

Enhanced the display_dot tcl procedure to call the quit function if such
is given and otherwise destroying the window.