Enhanced the draw_graph (and in C GR/GRl) to include line number of leaf nodes (optionally settable by vossrc variable INCLUDE-LINE_NBR-IN-GRAPH-DRAW).
Cleaned up the protect function to fix a potential bug with overlapping buffers.
Protect now can also return whether a change took place or not.

Cleaned up some anonymous names to make type error reporting more informative.
Also, enhanced the type error reporting by providing more informative error
messages (context and argument names). Still more can be done I suspect.

Enhanced the error reporting for functions that walk over fl structures.

Fixed some of the macros that had some left-over 32-bit assumptions.
Now it should be consistent (and all 64-bit).

Fixed a bug in the fl-tcl interface that made returning lists of strings
that had spaces in them not work.

Added a version of ENUM (in hfl) that takes a minimum size for the encoding
of the elements in the ENUM.

Added a function enum_value that given a name (as a string) for one
of the enumerated items, return its value. The return type determines
which ENUM value that is selected.

Added a UNION type.

Renamed the always_ff command to ff_block that just add
a (user selectable) flip-flop at the output of all the assignemnts.

Fixed (?) for the n-th time the w2bv function...

Enhanced the visulaization of symbolic values so that variables are shown,
rather than make all of them "S".

Modified the VIS command and added a simulate command that makes it
much easier to visualize + simulet a design. The VIS can now automatically
(by default) add the inputs and outputs to the waveform viewer.

Added a draw_RAM primitive.