Modifed the fclose function to return the status from the underlying fclose/pclose system call. NOTE: This can potentially break existing code!!!
Improved the error reporting when loading using -F flag so that
the error that causes the loading (and fl) to quit is now reported.

Improved implode to make sure the resulting string fits.

Added a system2 function call that is like system, except it returns a
triple of (return code, stdout, stderr), the latters a list of strings.

Fixed the new_yaccfl to at least work on one example.... First live version!