Improved the naming of temprary node names when parsing Verilog files. Now just the file name is used instead of the full path to the file.
Cleand up font settings when changing to a bigger/smaller font.

Modified STE_debug to start in a size that is 80% of the screen size.

Improved the node browser so that the nodes can be manually edited (e.g.,
if not full vector is to be displayed).

Fixed a typo in the "Start Editor" in the RTL viewer that made the system hang.

iCommented out the bifrost build at the moment until the ghc version issue
is resolved.